requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
gunicorn==19.9.0
idna==2.7
lxml==4.2.5
nltk==3.3
numpy==1.15.2
psycopg2==2.7.5
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
whitenoise==4.1
